Change priority of scc to the lowest
There is conflict with OCP default scc ((e.g., anyuid). For example,
if some of OCP pods were restarted it will by default select the
strict one with same priority. But the ingress nginx scc does not
allow run container with root user, which is not case for some of
OCP pods. Change the priority of scc to the lowest will fix this issue.